Tighten validator review contracts
Widen validator depth counters, document baseline rules, and keep banquet
starts from skipping active ceremony cleanup. Exercise the Warfront mutation
flag through the public transition path and preserve post-final banquet flows
with focused tests.

Move trace validator metadata to the top-level contract, add missing state graph
resource fields, attribute ceremony reward and honour events to real publishers,
and record deferred design decisions as ADRs.

+++ crates/skyjoust_stateright_validator/src/model.rs
@@ -12,7 +12,7 @@ pub type AlwaysProperty = fn(&SkyjoustInteractionModel, &SkyState) -> bool;
 pub struct SkyjoustInteractionModel {
     /// Exploration boundary. Increase for deeper interaction sequences, lower
     /// it for fast continuous integration smoke checks.
-    pub max_depth: u8,
+    pub max_depth: u16,
 }
 
 impl Default for SkyjoustInteractionModel {
